backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
numpy==1.26.4
pandas==2.2.3
fastembed==0.4.2
python-multipart==0.0.12
faiss-cpu==1.9.0
requests==2.32.3
python-dotenv==1.0.1
